% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file_format_matches.R
\name{check_file_format_matches}
\alias{check_file_format_matches}
\title{Check file format}
\usage{
check_file_format_matches(data_path, sys_path)
}
\arguments{
\item{data_path}{(char) A path or URL to a data file}

\item{sys_path}{(char) A path or URL to system metadata for the file}
}
\value{
(boolean) Whether or not the file type matches the formatId
}
\description{
This checks whether a file's format matches documentation
in a file's DataONE system metadata. The output of \verb{file filename.ext} (bash)
is checked against the \href{https://cn.dataone.org/cn/v2/formats}{formatId}
in the system metadata for that file.
}
\examples{
data_path <- "https://arcticdata.io/metacat/d1/mn/v2/object/urn:uuid:f4d6ffc1-9fc0-428e-ba7c-c99ff27922f3"
sys_path <- "https://arcticdata.io/metacat/d1/mn/v2/meta/urn:uuid:f4d6ffc1-9fc0-428e-ba7c-c99ff27922f3"
check_file_format_matches(data_path, sys_path)
}
