% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{check_text_file_format}
\alias{check_text_file_format}
\title{Example function}
\usage{
check_text_file_format(path, delimiter, header_lines)
}
\arguments{
\item{path}{(char) A path to a data file}

\item{delimiter}{(char) Record delimter to be used to parse the file}

\item{header_lines}{Number of header lines in the file}
}
\value{
(boolean) Whether or not the file could be parsed
}
\description{
Here is some documentation for this example function.
}
\examples{
path <- system.file("extdata/test.csv", package = "metadigRake")
check_text_file_format(path, ",", 1)
}
