% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_text_delimited_wellformed.R
\name{data_text_delimited_wellformed}
\alias{data_text_delimited_wellformed}
\title{Check if a text file is parsable}
\usage{
data_text_delimited_wellformed(path, delimiter, header_lines)
}
\arguments{
\item{path}{(char) A path to a data file}

\item{delimiter}{(char) Record delimter to be used to parse the file}

\item{header_lines}{Number of header lines in the file}
}
\value{
(boolean) Whether or not the file could be parsed
}
\description{
This is a wrapper around the helper \code{read_text_file} which returns TRUE if
the file is read in successfully and NULL if not.
}
\examples{
path <- system.file("extdata/test.csv", package = "metarake")
data_text_delimited_wellformed(path, ",", 1)
}
