% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_function}
\alias{example_function}
\title{Example function}
\usage{
example_function(msg)
}
\arguments{
\item{msg}{(string) A message to print}
}
\value{
A message printed to the console
}
\description{
Here is some documentation for this example function.
}
\examples{
example_function("Woohoo!")
}
