% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_pids.R
\name{get_data_pids}
\alias{get_data_pids}
\title{Get data pids}
\usage{
get_data_pids(pid, authoritative_node)
}
\arguments{
\item{pid}{(char) Identifier of a metadata document}

\item{authoritative_node}{(char) Authoritative member node identifier (eg: urn:node:ARCTIC)}
}
\value{
(char) List of data pids
}
\description{
Given a metadata pid and member node, return the data pids that are documented by
the metadata by querying the member node.
}
\examples{
pids <- get_data_pids("doi:10.18739/A2TH8BP4V", "urn:node:ARCTIC")
}
