% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mn_endpoint.R
\name{get_mn_endpoint}
\alias{get_mn_endpoint}
\title{Get member node endpoint}
\usage{
get_mn_endpoint(mn_id)
}
\arguments{
\item{mn_id}{(char) Identifier of a member node}
}
\value{
(char) Endpoint to query node
}
\description{
Given a member node identifier, return the endpoint.
}
\examples{
endpoint <- get_mn_endpoint("urn:node:ARCTIC")
}
