% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_system_metadata.R
\name{parse_system_metadata}
\alias{parse_system_metadata}
\title{Parse system metadata}
\usage{
parse_system_metadata(sysmeta)
}
\arguments{
\item{sysmeta}{(char) String representation of XML system metadata}
}
\value{
(SystemMetadata) SystemMetadata object
}
\description{
Helper to parse system metadata from a string in one function call
}
\examples{
sys_path <- system.file("extdata/sysmeta.xml", package = "metarake")
sys_char <- readChar(sys_path, file.info(sys_path)$size)
sys <- parse_system_metadata(sys_char)
}
