% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_text_file.R
\name{read_text_file}
\alias{read_text_file}
\title{Read tabular data from a delimited text file}
\usage{
read_text_file(path, delimiter, header_lines, ...)
}
\arguments{
\item{path}{(char) A path to a data file}

\item{delimiter}{(char) Record delimter to be used to parse the file}

\item{header_lines}{Number of header lines in the file}

\item{...}{other arguments to \code{read.table}}
}
\value{
(df) A parsed data.frame
}
\description{
This function takes a path, delimiter, and header lines to read in a file
returning the data as a data.frame if it is successful, and NULL if not.
}
\examples{
path <- system.file("extdata/test.csv", package = "metadigRake")
read_text_file(path, ",", 1)
}
